Mixnode keys loaded in run command
